feat: IAM gateway handlers + map-based dispatcher
Add IAM request dispatcher with generic iamHandler pattern matching
EC2's existing map-based dispatch. Thin gateway handlers validate
inputs before delegating to the IAM service layer.

- gateway/iam.go: IAMHandler type, iamHandler generic, iamActions map,
  IAM_Request dispatcher (8 actions)
- gateway/iam/user.go: CreateUser, GetUser, ListUsers, DeleteUser
- gateway/iam/key.go: CreateAccessKey, ListAccessKeys, DeleteAccessKey,
  UpdateAccessKey
- gateway/gateway.go: IAMService + IAMMasterKey fields on GatewayConfig,
  GenerateIAMErrorResponse with IAM-style XML format, ErrorHandler
  dispatches IAM vs EC2 error XML
- Fix gosec G115: change getOrCreateBucket history param to uint8

+++ hive/handlers/iam/service_impl.go
@@ -63,10 +63,10 @@ func NewIAMServiceImpl(natsConn *nats.Conn, masterKey []byte) (*IAMServiceImpl, 
 	}, nil
 }
 
-func getOrCreateBucket(js nats.JetStreamContext, name string, history int) (nats.KeyValue, error) {
+func getOrCreateBucket(js nats.JetStreamContext, name string, history uint8) (nats.KeyValue, error) {
 	kv, err := js.CreateKeyValue(&nats.KeyValueConfig{
 		Bucket:  name,
-		History: uint8(history),
+		History: history,
 	})
 	if err != nil {
 		kv, err = js.KeyValue(name)
